API documentation

API documentation

Validation

Analysis

Text corpora

Exported types and functions

Internals

Indexing

Named entity management and topic modelling

Exported types and functions

Internals

Validation

[ ] validate cataloging and citation structure of all texts

[ ] validate orthography

[ ] validate syntactic validity and completeness of DSE indexing

[ ] validate syntactic validity and referential integrity of scholia references

[ ] validate syntactic validity and referential integrity of persName elements

[ ] validate syntactic validity and referential integrity of placeName elements

[ ] validate syntactic validity and referential integrity of rs elements

Analysis

Text corpora

Exported types and functions

HmtArchive.Archive — Type

The HMT archive in a local file system.

HmtArchive.corpus — Function

Builds a single corpus comprising Venetus A Iliad and scholia in mulitvalent archival XML edition, and univocal diplomatic and normalized editions.

corpus(archive)

HmtArchive.iliadxmlcorpus — Function

Compile a single citable edition of Venetus A Iliad.

iliadxmlcorpus(archive)

HmtArchive.scholiaxmlcorpus — Function

Compile a single citable edition of all Venetus A scholia.

scholiaxmlcorpus(archive)

HmtArchive.iliaddipl — Function

Compose a diplomatic edition of Iliad texts.

iliaddipl(archive)

HmtArchive.iliadnormed — Function

Compose a normalized edition of Iliad texts.

iliadnormed(archive)

HmtArchive.scholiadipl — Function

Compose a diplomatic edition of scholia.

scholiadipl(archive)

HmtArchive.scholianormed — Function

Compose a normalized edition of scholia.

scholianormed(archive)

Internals

HmtArchive.iliaddir — Function

Directory with archival XML for Iliad editions.

iliaddir(archive)

HmtArchive.scholiadir — Function

Find directory with archival XML source for scholia editions.

scholiadir(archive)

HmtArchive.scholiaforbookdoc — Function

Make a CitableCorpus for one scholia document in one book.

scholiaforbookdoc(docroot, bk)

Arguments

docroot is the parsed XML root node for one scholia document such as "Venetus A main scholia" or "Venetus A intermarginal scholia"

bk is the Iliad book number you want.

Indexing

HmtArchive.indexscholia — Function

Compose a list of Tuples pairing canonical CtsUrns for scholia with CtsUrns for the Iliad text.

indexscholia(archive)

HmtArchive.indexrefnode — Function

Index a scholion node to an Iliad passage.

(

indexrefnode(cn)

)

Arguments

cn should be the citable node for the ref node of a scholion record.

The function returns a pairing of the canonical reference for the scholion with the Iliad Urn in the ref node's text content.

Named entity management and topic modelling

Exported types and functions

HmtArchive.TMEditionBuilder — Type

Builder to compose editions for topic modelling.

HmtArchive.editednode — Function

Builder for constructing a citable node for a diplomatic text from a citable node in archival XML.

editednode(builder, citablenode)

HmtArchive.expandabbr — Function

Expand abbreviated URN string to full Cite2Urn using neregistry.

expandabbr(s)

Expand abbreviated URN to full Cite2Urn using neregistry.

expandabbr(abbr)

Internals

HmtArchive.editedTMelement — Function

Compose edited text of a given XML element using a given builder.

editedTMelement(el, accum)

Piggy back on the MidNormalizedBuilder to normalize everything except named entities.

HmtArchive.editedTMtext — Function

Convert a parsed node of XML to appropriate text string.

editedTMtext(n)
editedTMtext(n, accum)

HmtArchive.isnamedentity — Function

True if elname is the name of an element tagging a named entity.

isnamedentity(elname)

Currently only supports persName and placeName. HMT usage of rs should be supported by consulting type attribute, since there are uses of TEI rs for both named entities and other annotations that are only distinguished by the type attribute

HmtArchive.neregistry — Function

Create dictionary mapping abbreviated collection names to full Cite2Urns.

neregistry()