Published releases: API documentation

Loading and identifing a CEX release

HmtArchive.Analysis.hmt_cex — Function

Download current HMT release in CEX format.

hmt_cex()

HmtArchive.Analysis.hmt_releaseinfo — Function

Extract release info from library header of cexsrc.

hmt_releaseinfo(cexsrc; delimiter)

Instantiating parts of the release

HmtArchive.Analysis.hmt_images — Function

Load HMT image collections from src.

hmt_images(src; delimiter)

Load HMT image collections from current CEX release of HMT project.

hmt_images()

HmtArchive.Analysis.hmt_codices — Function

Load HMT codex page collections from src.

hmt_codices(src; delimiter)

Load HMT codex page collections from current CEX release of HMT project.

hmt_codices()

HmtArchive.Analysis.hmt_textcatalog — Function

Load HMT text catalog from src.

hmt_textcatalog(src; delimiter)

Load HMT text catalog from current CEX release of HMT project.

hmt_textcatalog()

HmtArchive.Analysis.hmt_diplomatic — Function

Load diplomatic corpus from src.

hmt_diplomatic(src; delimiter)

Load diplomatic corpus from current CEX release of HMT project.

hmt_diplomatic()

HmtArchive.Analysis.hmt_normalized — Function

Load normalized corpus from src.

hmt_normalized(src; delimiter)

Load normalized corpus from current CEX release of HMT project.

hmt_normalized()

HmtArchive.Analysis.hmt_dse — Function

Load DSE records from src.

hmt_dse(src; delimiter)

Load DSE records from current CEX release of HMT project.

hmt_dse()

Missing docstring.

Missing docstring for hmt_commentary. Check Documenter's build log for details.

HmtArchive.Analysis.hmt_pageindex — Function

Load HMT index of Iliad lines to pages from src.

hmt_pageindex(src; delimiter)

Load HMT index of Iliad lines to pages from current CEX release of HMT project.

hmt_pageindex()

HmtArchive.Analysis.hmt_paragraphs — Function

Load HMT data indexing chunking of texts in paragraphs from src.

hmt_paragraphs(src; delimiter)

Load HMT data indexing chunking of texts in paragraphs from current CEX release of HMT project.

hmt_paragraphs()

HmtArchive.Analysis.hmt_persnames — Function

Load HMT authority list for personal names from src.

hmt_persnames(src; delimiter)

Load HMT authority list for personal names from current release.

hmt_persnames()

HmtArchive.Analysis.hmt_placenames — Function

Load HMT authority list for place names from src.

hmt_placenames(src; delimiter)

Load HMT authority list for place names from current release.

hmt_placenames()

Tables summarizing parts of the release

HmtArchive.Analysis.coltbl_pagecounts — Function

Load current release and compose a table of page counts per codex.

coltbl_pagecounts()

Compose a table of page counts per codex in the CEX source src.

coltbl_pagecounts(src)

Compose a table of page counts per codex in codd.

coltbl_pagecounts(codd)

HmtArchive.Analysis.coltbl_imagecounts — Function

Compose a table of image counts per image collection in codd.

coltbl_imagecounts(imgs)

Compose a table of image counts per image collection in src.

coltbl_imagecounts(src)

Load current release and compose a table of image counts per image collection.

coltbl_imagecounts()